LayoutTests/accessibility/aria-modal-expected.txt

This tests that aria-modal on dialog makes other elements inert.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Dialog is displaying
PASS backgroundAccessible() is false
Dialog is not displaying
PASS backgroundAccessible() is true
Dialog is displaying
PASS backgroundAccessible() is false
PASS okButton.isIgnored is false
Dialog is not displaying and aria-modal=true
PASS backgroundAccessible() is true
Dialog is displaying
PASS backgroundAccessible() is false
Dialog is displaying and aria-hidden=true
PASS backgroundAccessible() is true
Dialog is displaying and aria-hidden=false
PASS backgroundAccessible() is false
Dialog is not displaying with opacity 0
PASS backgroundAccessible() is true
Dialog is displaying with opacity 1
PASS backgroundAccessible() is false
Dialog is not displaying with parent opacity 0
PASS backgroundAccessible() is true
Dialog is displaying with parent opacity .5
PASS backgroundAccessible() is false
Dialog is removed from DOM
PASS backgroundAccessible() is true
PASS successfullyParsed is true

TEST COMPLETE
Other page content with a dummy focusable element

Display a dialog
